+++ CHANGELOG.md
@@ -6,10 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Default client level message parser error handling hook.
 - Component arguments to the relevante context create message methods.
 
 ### Changed
 - Bumped minimum Hikari version to 2.0.0.dev102.
+
+### Fixed
+- SlashContext.edit_last_response can now be used to edit a deferred response.
 
 ### Removed
 - suppress_exceptions from Client.dispatch_client_callback cause it was half-arsed and didn't make sense.